Started Building GUI
The background .png has been set up, along with the Feedback Rotary Slider, and the two time delay rotary sliders for the left and right channels. The knobs can turn and modify the internal variables of their respective modules. The values themselves are smoothed to prevent clipping and distortion. The sliders are attached to the AudioProcessorValueTreeState, in hopes of working modifications inside of DAWs. Still to implement is the saving of data when the plugin is closed and reopened. Still to implement is the filters and the mixers.