UI finished
Implemented the remainder of the knobs. Added the bypass switch and the mono switch. I'm still missing the button to link the two delay timers, but I'm not sure if I want to implement that. Something that I might change is that the delay time reaches up to 4000ms, which is 4 seconds long. Musically might still be fine but the most usable delays only go up to about 300ms. The usable range of the delay knob is only about 15% of the entire range. So I might export the plugin, then test it on FL Studio and see how it feels to use the entire 4000ms. IF it doesn't seem that useable i'll knock the range down. Still, I need to finish the link button